[ds-singly-ll] break down into header, functions, main files
