+++ singly_linked_list.c
@@ -1,28 +1,4 @@
-//ADT implementation: https://runestone.academy/ns/books/published/pythonds3/BasicDS/TheUnorderedListAbstractDataType.html
-
-#include <stdio.h>
-#include <stdlib.h>
-
-struct Node{
-    int data;
-    struct Node* next;
-};
-
-typedef struct
-{
-    struct Node* head;
-} List;
-
-List list_create(); // returns empty list
-void list_free(List* self); // deallocates list
-void list_push_front(List* self, int value);
-int list_pop_front(List* self);
-struct Node* list_find(List* self, int value); // NULL if not found
-void list_erase_after(List* self, struct Node* node);
-void list_push_back(List* self, int value);
-void list_merge_to_back(List* self, List* other);
-void list_merge_to_front(List* self, List* other);
-void list_splice_after(List* self, struct Node* after, List* other);
+#include "singly_linked_list.h"
 
 
 void list_splice_after(List* self, struct Node* after, List* other) {
@@ -194,48 +170,4 @@ List list_create(void) {
     list.head = NULL;
 
     return list;
-}
-
-int main() {
-    /*
-    List list = list_create();
-    list_push_back(&list, 1);
-    list_free(&list);
-    assert(list.head == NULL);
-    */
-
-    List list = list_create();
-    List other_list = list_create();
-
-    list_push_front(&list, 3);
-    list_push_front(&list, 2);
-    list_push_front(&list, 1);
-    list_print(&list);
-
-    list_push_front(&other_list, 6);
-    list_push_front(&other_list, 5);
-    list_push_front(&other_list, 4);
-    list_print(&other_list);
-
-    struct Node* p_found_node_value = list_find(&list, 1);
-
-    list_splice_after(&list, p_found_node_value, &other_list);
-    list_print(&list);
-
-    list_merge_to_back(&list, &other_list);
-    list_print(&list);
-
-    list_merge_to_front(&list, &other_list);
-    list_print(&list);
-
-    list_erase_after(&list, p_found_node_value);
-    list_print(&list);
-
-    int i_val_1 = list_pop_front(&list);
-    int i_val_2 = list_pop_front(&list);
-    int i_val_3 = list_pop_front(&list);
-
-    list_free(&list);
-
-    return 0;
 }
